chore: remove legacy engine file
