+++ src/core/fsm/execution_engine/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+"""FSM execution engine package."""
+
+from .core import FSMCore
+
+FSMCoreV2 = FSMCore
+FiniteStateMachine = FSMCore
+WorkflowEngine = FSMCore
+
+__all__ = ["FSMCore", "FSMCoreV2", "FiniteStateMachine", "WorkflowEngine"]
