+++ src/core/fsm/execution_engine/common.py
@@ -1,0 +1,54 @@
+#!/usr/bin/env python3
+"""Shared imports for FSM execution engine modules."""
+
+try:  # pragma: no cover - support standalone execution
+    from ..helpers import load_fsm_config
+    from ..types import FSMConfig
+    from ..state_machine import (
+        StateDefinition,
+        TransitionDefinition,
+        WorkflowInstance,
+        StateExecutionResult,
+        StateStatus,
+        TransitionType,
+        WorkflowPriority,
+        StateHandler,
+        TransitionHandler,
+    )
+    from ..utils import can_execute_transition, build_context, execute_actions
+except Exception:  # pragma: no cover - fallback imports
+    from helpers import load_fsm_config  # type: ignore
+    from types import FSMConfig  # type: ignore
+    from state_machine import (  # type: ignore
+        StateDefinition,
+        TransitionDefinition,
+        WorkflowInstance,
+        StateExecutionResult,
+        StateStatus,
+        TransitionType,
+        WorkflowPriority,
+        StateHandler,
+        TransitionHandler,
+    )
+    from utils import (  # type: ignore
+        can_execute_transition,
+        build_context,
+        execute_actions,
+    )
+
+__all__ = [
+    "load_fsm_config",
+    "FSMConfig",
+    "StateDefinition",
+    "TransitionDefinition",
+    "WorkflowInstance",
+    "StateExecutionResult",
+    "StateStatus",
+    "TransitionType",
+    "WorkflowPriority",
+    "StateHandler",
+    "TransitionHandler",
+    "can_execute_transition",
+    "build_context",
+    "execute_actions",
+]
