+++ src/core/fsm/models.py
@@ -16,41 +16,27 @@ from enum import Enum
 from typing import Dict, List, Optional, Any
 from datetime import datetime
 
+from .state_machine import (
+    StateStatus,
+    TransitionType,
+    WorkflowPriority,
+    StateDefinition,
+    TransitionDefinition,
+    WorkflowInstance,
+    StateExecutionResult,
+    StateHandler,
+    TransitionHandler,
+)
+
 
 # ============================================================================
-# CORE FSM ENUMS
+# TASK MANAGEMENT ENUMS
 # ============================================================================
-
-class StateStatus(Enum):
-    """State execution status."""
-    PENDING = "pending"
-    ACTIVE = "active"
-    COMPLETED = "completed"
-    FAILED = "failed"
-    SKIPPED = "skipped"
-    TIMEOUT = "timeout"
-
-
-class TransitionType(Enum):
-    """Types of state transitions."""
-    AUTOMATIC = "automatic"
-    MANUAL = "manual"
-    CONDITIONAL = "conditional"
-    TIMEOUT = "timeout"
-    ERROR = "error"
-
-
-class WorkflowPriority(Enum):
-    """Workflow priority levels."""
-    LOW = "low"
-    NORMAL = "normal"
-    HIGH = "high"
-    CRITICAL = "critical"
-    EMERGENCY = "emergency"
 
 
 class TaskState(Enum):
     """Task state enumeration."""
+
     NEW = "new"
     ONBOARDING = "onboarding"
     IN_PROGRESS = "in_progress"
@@ -62,6 +48,7 @@ class TaskState(Enum):
 
 class TaskPriority(Enum):
     """Task priority enumeration."""
+
     LOW = "low"
     NORMAL = "normal"
     HIGH = "high"
@@ -70,6 +57,7 @@ class TaskPriority(Enum):
 
 class BridgeState(Enum):
     """Bridge state enumeration."""
+
     IDLE = "idle"
     CONNECTING = "connecting"
     CONNECTED = "connected"
@@ -78,72 +66,14 @@ class BridgeState(Enum):
 
 
 # ============================================================================
-# CORE FSM DATA MODELS
-# ============================================================================
-
-@dataclass
-class StateDefinition:
-    """State definition structure."""
-    name: str
-    description: str
-    entry_actions: List[str]
-    exit_actions: List[str]
-    timeout_seconds: Optional[float]
-    retry_count: int
-    retry_delay: float
-    required_resources: List[str]
-    dependencies: List[str]
-    metadata: Dict[str, Any]
-
-
-@dataclass
-class TransitionDefinition:
-    """Transition definition structure."""
-    from_state: str
-    to_state: str
-    transition_type: TransitionType
-    condition: Optional[str]
-    priority: int
-    timeout_seconds: Optional[float]
-    actions: List[str]
-    metadata: Dict[str, Any]
-
-
-@dataclass
-class WorkflowInstance:
-    """Workflow instance tracking."""
-    workflow_id: str
-    workflow_name: str
-    current_state: str
-    state_history: List[Dict[str, Any]]
-    start_time: datetime
-    last_update: datetime
-    status: StateStatus
-    priority: WorkflowPriority
-    metadata: Dict[str, Any]
-    error_count: int
-    retry_count: int
-
-
-@dataclass
-class StateExecutionResult:
-    """Result of state execution."""
-    state_name: str
-    execution_time: float
-    status: StateStatus
-    output: Dict[str, Any]
-    error_message: Optional[str]
-    metadata: Dict[str, Any]
-    timestamp: datetime
-
-
-# ============================================================================
 # TASK MANAGEMENT DATA MODELS
 # ============================================================================
+
 
 @dataclass
 class FSMTask:
     """FSM task data structure."""
+
     id: str
     title: str
     description: str
@@ -163,11 +93,13 @@ class FSMTask:
 
     def add_evidence(self, agent_id: str, evidence: Dict[str, Any]):
         """Add evidence to the task."""
-        self.evidence.append({
-            "timestamp": datetime.now().isoformat(),
-            "agent_id": agent_id,
-            "evidence": evidence,
-        })
+        self.evidence.append(
+            {
+                "timestamp": datetime.now().isoformat(),
+                "agent_id": agent_id,
+                "evidence": evidence,
+            }
+        )
         self.updated_at = datetime.now().isoformat()
 
     def update_state(self, new_state: TaskState):
@@ -187,7 +119,7 @@ class FSMTask:
             "created_at": self.created_at,
             "updated_at": self.updated_at,
             "evidence": self.evidence,
-            "metadata": self.metadata
+            "metadata": self.metadata,
         }
         return task_dict
 
@@ -202,6 +134,7 @@ class FSMTask:
 @dataclass
 class FSMUpdate:
     """FSM update message structure."""
+
     update_id: str
     task_id: str
     agent_id: str
@@ -214,6 +147,7 @@ class FSMUpdate:
 @dataclass
 class FSMCommunicationEvent:
     """FSM communication event structure."""
+
     event_id: str
     event_type: str
     source_agent: str
@@ -221,32 +155,3 @@ class FSMCommunicationEvent:
     message: str
     timestamp: str
     metadata: Dict[str, Any] = None
-
-
-# ============================================================================
-# ABSTRACT BASE CLASSES
-# ============================================================================
-
-class StateHandler:
-    """Abstract base class for state handlers."""
-    
-    def execute(self, context: Dict[str, Any]) -> StateExecutionResult:
-        """Execute the state logic."""
-        raise NotImplementedError
-    
-    def can_transition(self, context: Dict[str, Any]) -> bool:
-        """Check if transition to this state is allowed."""
-        raise NotImplementedError
-
-
-class TransitionHandler:
-    """Abstract base class for transition handlers."""
-    
-    def evaluate(self, context: Dict[str, Any]) -> bool:
-        """Evaluate transition condition."""
-        raise NotImplementedError
-    
-    def execute(self, context: Dict[str, Any]) -> Dict[str, Any]:
-        """Execute transition actions."""
-        raise NotImplementedError
-
